Always convert size matcher to size_t.

+++ include/mcga/matchers/iterable.hpp
@@ -143,7 +143,7 @@ constexpr auto hasSize(const T& size) {
     if constexpr (Matcher<T>) {
         return IterableSizeMatcher<T>(size);
     } else {
-        return IterableSizeMatcher(isEqualTo(size));
+        return IterableSizeMatcher(isEqualTo(static_cast<std::size_t>(size)));
     }
 }
